Refactor authentication logic extracting common parts
