+++ android/java/org/chromium/chrome/browser/crypto_wallet/fragments/UnlockWalletFragment.java
@@ -5,18 +5,9 @@
 
 package org.chromium.chrome.browser.crypto_wallet.fragments;
 
-import static android.hardware.biometrics.BiometricPrompt.BIOMETRIC_ERROR_USER_CANCELED;
-
-import android.annotation.SuppressLint;
-import android.content.Context;
-import android.hardware.biometrics.BiometricPrompt;
-import android.hardware.biometrics.BiometricPrompt.AuthenticationCallback;
-import android.hardware.biometrics.BiometricPrompt.AuthenticationResult;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.CancellationSignal;
 import android.text.Editable;
-import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -26,9 +17,7 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 import androidx.appcompat.widget.AppCompatButton;
-import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentActivity;
 
 import com.google.android.material.textfield.TextInputEditText;
@@ -40,22 +29,9 @@ import org.chromium.chrome.browser.app.helpers.Api33AndPlusBackPressHelper;
 import org.chromium.chrome.browser.crypto_wallet.listeners.OnNextPage;
 import org.chromium.chrome.browser.crypto_wallet.util.KeystoreHelper;
 import org.chromium.chrome.browser.crypto_wallet.util.Utils;
-import org.chromium.ui.widget.Toast;
 
-import java.io.IOException;
-import java.security.InvalidAlgorithmParameterException;
-import java.security.InvalidKeyException;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.UnrecoverableEntryException;
-import java.security.cert.CertificateException;
-import java.util.concurrent.Executor;
-
-import javax.crypto.BadPaddingException;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.NoSuchPaddingException;
-
-public class UnlockWalletFragment extends BaseWalletNextPageFragment {
+public class UnlockWalletFragment extends BaseWalletNextPageFragment
+        implements BaseWalletNextPageFragment.BiometricAuthenticationCallback {
 
     private TextInputEditText mUnlockWalletPassword;
     private TextInputLayout mUnlockWalletPasswordLayout;
@@ -146,7 +122,7 @@ public class UnlockWalletFragment extends BaseWalletNextPageFragment {
                 v -> {
                     if (Utils.isBiometricSupported(requireContext())) {
                         // noinspection NewApi
-                        showBiometricAuthenticationDialog();
+                        showBiometricAuthenticationDialog(mBiometricUnlockButton, this);
                     }
                 });
 
@@ -155,7 +131,7 @@ public class UnlockWalletFragment extends BaseWalletNextPageFragment {
 
             mBiometricUnlockButton.setVisibility(View.VISIBLE);
             // noinspection NewApi
-            showBiometricAuthenticationDialog();
+            showBiometricAuthenticationDialog(mBiometricUnlockButton, this);
         }
     }
 
@@ -168,88 +144,90 @@ public class UnlockWalletFragment extends BaseWalletNextPageFragment {
         }
     }
 
-    @SuppressLint("MissingPermission")
-    @RequiresApi(api = Build.VERSION_CODES.P)
-    private void showBiometricAuthenticationDialog() {
-        final AuthenticationCallback authenticationCallback =
-                new AuthenticationCallback() {
-                    @Override
-                    public void onAuthenticationSucceeded(AuthenticationResult result) {
-                        super.onAuthenticationSucceeded(result);
-
-                        final KeyringService keyringService = getKeyringService();
-                        String unlockWalletPassword;
-
-                        try {
-                            unlockWalletPassword = KeystoreHelper.decryptText();
-                        } catch (InvalidAlgorithmParameterException
-                                | UnrecoverableEntryException
-                                | NoSuchPaddingException
-                                | IllegalBlockSizeException
-                                | CertificateException
-                                | KeyStoreException
-                                | NoSuchAlgorithmException
-                                | BadPaddingException
-                                | IOException
-                                | InvalidKeyException e) {
-                            // KeystoreHelper.decryptText() may throw a long list
-                            // of exceptions.
-                            showBiometricAuthenticationButton();
-                            return;
-                        }
-
-                        if (TextUtils.isEmpty(unlockWalletPassword) || keyringService == null) {
-                            showBiometricAuthenticationButton();
-                            return;
-                        }
-
-                        keyringService.unlock(
-                                unlockWalletPassword,
-                                unlockResult -> {
-                                    if (unlockResult) {
-                                        if (mOnNextPage != null) {
-                                            mOnNextPage.showWallet(false);
-                                        }
-                                    } else {
-                                        showBiometricAuthenticationButton();
-                                        mUnlockWalletPassword.setError(
-                                                getString(R.string.incorrect_password_error));
-                                    }
-                                });
-                    }
-
-                    @Override
-                    public void onAuthenticationError(int errorCode, CharSequence errString) {
-                        super.onAuthenticationError(errorCode, errString);
-
-                        final Context context = getContext();
-                        if (!TextUtils.isEmpty(errString) && context != null) {
-                            Toast.makeText(context, errString, Toast.LENGTH_SHORT).show();
-                        }
-                        showBiometricAuthenticationButton();
-                    }
-                };
-        Executor executor = ContextCompat.getMainExecutor(requireContext());
-        new BiometricPrompt.Builder(requireContext())
-                .setTitle(getResources().getString(R.string.fingerprint_unlock))
-                .setDescription(getResources().getString(R.string.use_fingerprint_text))
-                .setNegativeButton(
-                        getResources().getString(android.R.string.cancel),
-                        executor,
-                        (dialog, which) ->
-                                authenticationCallback.onAuthenticationError(
-                                        BIOMETRIC_ERROR_USER_CANCELED, ""))
-                .build()
-                .authenticate(new CancellationSignal(), executor, authenticationCallback);
-    }
-
-    /** Shows biometric authentication button if supported and if it was previously set. */
-    private void showBiometricAuthenticationButton() {
-        if (Utils.isBiometricSupported(requireContext())
-                && KeystoreHelper.shouldUseBiometricToUnlock()) {
-            mBiometricUnlockButton.setVisibility(View.VISIBLE);
-        } else {
-            mBiometricUnlockButton.setVisibility(View.GONE);
+    @Override
+    public void authenticationSuccess(@NonNull String unlockWalletPassword) {
+        Utils.clearClipboard(unlockWalletPassword);
+        mUnlockWalletPassword.setText(null);
+        if (mOnNextPage != null) {
+            mOnNextPage.showWallet(false);
         }
     }
+
+    //    @SuppressLint("MissingPermission")
+    //    @RequiresApi(api = Build.VERSION_CODES.P)
+    //    private void showBiometricAuthenticationDialog() {
+    //        final AuthenticationCallback authenticationCallback =
+    //                new AuthenticationCallback() {
+    //                    @Override
+    //                    public void onAuthenticationSucceeded(
+    //                            AuthenticationResult authenticationResult) {
+    //                        super.onAuthenticationSucceeded(authenticationResult);
+    //
+    //                        final KeyringService keyringService = getKeyringService();
+    //                        String unlockWalletPassword;
+    //
+    //                        try {
+    //                            unlockWalletPassword = KeystoreHelper.decryptText();
+    //                        } catch (InvalidAlgorithmParameterException
+    //                                | UnrecoverableEntryException
+    //                                | NoSuchPaddingException
+    //                                | IllegalBlockSizeException
+    //                                | CertificateException
+    //                                | KeyStoreException
+    //                                | NoSuchAlgorithmException
+    //                                | BadPaddingException
+    //                                | IOException
+    //                                | InvalidKeyException e) {
+    //                            // KeystoreHelper.decryptText() may throw a long list
+    //                            // of exceptions.
+    //                            showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                            return;
+    //                        }
+    //
+    //                        if (TextUtils.isEmpty(unlockWalletPassword) || keyringService == null)
+    // {
+    //                            showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                            return;
+    //                        }
+    //
+    //                        keyringService.unlock(
+    //                                unlockWalletPassword,
+    //                                result -> {
+    //                                    if (result) {
+    //                                        Utils.clearClipboard(unlockWalletPassword);
+    //                                        mUnlockWalletPassword.setText(null);
+    //                                        if (mOnNextPage != null) {
+    //                                            mOnNextPage.showWallet();
+    //                                        }
+    //                                    } else {
+    //
+    // showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                                    }
+    //                                });
+    //                    }
+    //
+    //                    @Override
+    //                    public void onAuthenticationError(int errorCode, CharSequence errString) {
+    //                        super.onAuthenticationError(errorCode, errString);
+    //
+    //                        final Context context = getContext();
+    //                        if (!TextUtils.isEmpty(errString) && context != null) {
+    //                            Toast.makeText(context, errString, Toast.LENGTH_SHORT).show();
+    //                        }
+    //                        showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                    }
+    //                };
+    //        Executor executor = ContextCompat.getMainExecutor(requireContext());
+    //        new Builder(requireContext())
+    //                .setTitle(getResources().getString(R.string.fingerprint_unlock))
+    //                .setDescription(getResources().getString(R.string.use_fingerprint_text))
+    //                .setNegativeButton(
+    //                        getResources().getString(android.R.string.cancel),
+    //                        executor,
+    //                        (dialog, which) ->
+    //                                authenticationCallback.onAuthenticationError(
+    //                                        BIOMETRIC_ERROR_USER_CANCELED, ""))
+    //                .build()
+    //                .authenticate(new CancellationSignal(), executor, authenticationCallback);
+    //    }
 }
