+++ android/java/org/chromium/chrome/browser/crypto_wallet/fragments/onboarding/OnboardingBackupWalletFragment.java
@@ -5,16 +5,9 @@
 
 package org.chromium.chrome.browser.crypto_wallet.fragments.onboarding;
 
-import static android.hardware.biometrics.BiometricPrompt.BIOMETRIC_ERROR_USER_CANCELED;
-
-import android.annotation.SuppressLint;
-import android.content.Context;
-import android.hardware.biometrics.BiometricPrompt;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.CancellationSignal;
 import android.text.Editable;
-import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -23,9 +16,7 @@ import android.widget.ImageView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 import androidx.appcompat.widget.AppCompatButton;
-import androidx.core.content.ContextCompat;
 import androidx.fragment.app.FragmentActivity;
 
 import com.google.android.material.textfield.TextInputEditText;
@@ -34,25 +25,13 @@ import com.google.android.material.textfield.TextInputLayout;
 import org.chromium.brave_wallet.mojom.KeyringService;
 import org.chromium.chrome.R;
 import org.chromium.chrome.browser.app.helpers.Api33AndPlusBackPressHelper;
+import org.chromium.chrome.browser.crypto_wallet.fragments.BaseWalletNextPageFragment;
 import org.chromium.chrome.browser.crypto_wallet.listeners.OnNextPage;
 import org.chromium.chrome.browser.crypto_wallet.util.KeystoreHelper;
 import org.chromium.chrome.browser.crypto_wallet.util.Utils;
-import org.chromium.ui.widget.Toast;
 
-import java.io.IOException;
-import java.security.InvalidAlgorithmParameterException;
-import java.security.InvalidKeyException;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.UnrecoverableEntryException;
-import java.security.cert.CertificateException;
-import java.util.concurrent.Executor;
-
-import javax.crypto.BadPaddingException;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.NoSuchPaddingException;
-
-public class OnboardingBackupWalletFragment extends BaseOnboardingWalletFragment {
+public class OnboardingBackupWalletFragment extends BaseOnboardingWalletFragment
+        implements BaseWalletNextPageFragment.BiometricAuthenticationCallback {
     private TextInputEditText mUnlockWalletPassword;
     private TextInputLayout mUnlockWalletPasswordLayout;
     private AppCompatButton mUnlockButton;
@@ -142,7 +121,7 @@ public class OnboardingBackupWalletFragment extends BaseOnboardingWalletFragment
                 v -> {
                     if (Utils.isBiometricSupported(requireContext())) {
                         // noinspection NewApi
-                        showBiometricAuthenticationDialog();
+                        showBiometricAuthenticationDialog(mBiometricUnlockButton, this);
                     }
                 });
 
@@ -151,7 +130,7 @@ public class OnboardingBackupWalletFragment extends BaseOnboardingWalletFragment
 
             mBiometricUnlockButton.setVisibility(View.VISIBLE);
             // noinspection NewApi
-            showBiometricAuthenticationDialog();
+            showBiometricAuthenticationDialog(mBiometricUnlockButton, this);
         }
     }
 
@@ -160,91 +139,93 @@ public class OnboardingBackupWalletFragment extends BaseOnboardingWalletFragment
         return false;
     }
 
-    @SuppressLint("MissingPermission")
-    @RequiresApi(api = Build.VERSION_CODES.P)
-    private void showBiometricAuthenticationDialog() {
-        final BiometricPrompt.AuthenticationCallback authenticationCallback =
-                new BiometricPrompt.AuthenticationCallback() {
-                    @Override
-                    public void onAuthenticationSucceeded(
-                            BiometricPrompt.AuthenticationResult authenticationResult) {
-                        super.onAuthenticationSucceeded(authenticationResult);
-
-                        final KeyringService keyringService = getKeyringService();
-                        String unlockWalletPassword;
-
-                        try {
-                            unlockWalletPassword = KeystoreHelper.decryptText();
-                        } catch (InvalidAlgorithmParameterException
-                                | UnrecoverableEntryException
-                                | NoSuchPaddingException
-                                | IllegalBlockSizeException
-                                | CertificateException
-                                | KeyStoreException
-                                | NoSuchAlgorithmException
-                                | BadPaddingException
-                                | IOException
-                                | InvalidKeyException e) {
-                            // KeystoreHelper.decryptText() may throw a long list
-                            // of exceptions.
-                            showBiometricAuthenticationButton();
-                            return;
-                        }
-
-                        if (TextUtils.isEmpty(unlockWalletPassword) || keyringService == null) {
-                            showBiometricAuthenticationButton();
-                            return;
-                        }
-
-                        keyringService.unlock(
-                                unlockWalletPassword,
-                                result -> {
-                                    if (result) {
-                                        Utils.clearClipboard(unlockWalletPassword);
-                                        mUnlockWalletPassword.setText(null);
-                                        mOnboardingViewModel.setPassword(unlockWalletPassword);
-                                        if (mOnNextPage != null) {
-                                            mOnNextPage.incrementPages(1);
-                                        }
-                                    } else {
-                                        mUnlockWalletPasswordLayout.setError(
-                                                getString(R.string.incorrect_password_error));
-                                    }
-                                });
-                    }
-
-                    @Override
-                    public void onAuthenticationError(int errorCode, CharSequence errString) {
-                        super.onAuthenticationError(errorCode, errString);
-
-                        final Context context = getContext();
-                        if (!TextUtils.isEmpty(errString) && context != null) {
-                            Toast.makeText(context, errString, Toast.LENGTH_SHORT).show();
-                        }
-                        showBiometricAuthenticationButton();
-                    }
-                };
-        Executor executor = ContextCompat.getMainExecutor(requireContext());
-        new BiometricPrompt.Builder(requireContext())
-                .setTitle(getResources().getString(R.string.fingerprint_unlock))
-                .setDescription(getResources().getString(R.string.use_fingerprint_text))
-                .setNegativeButton(
-                        getResources().getString(android.R.string.cancel),
-                        executor,
-                        (dialog, which) ->
-                                authenticationCallback.onAuthenticationError(
-                                        BIOMETRIC_ERROR_USER_CANCELED, ""))
-                .build()
-                .authenticate(new CancellationSignal(), executor, authenticationCallback);
-    }
-
-    /** Shows biometric authentication button if supported and if it was previously set. */
-    private void showBiometricAuthenticationButton() {
-        if (Utils.isBiometricSupported(requireContext())
-                && KeystoreHelper.shouldUseBiometricToUnlock()) {
-            mBiometricUnlockButton.setVisibility(View.VISIBLE);
-        } else {
-            mBiometricUnlockButton.setVisibility(View.GONE);
+    @Override
+    public void authenticationSuccess(@NonNull String unlockWalletPassword) {
+        Utils.clearClipboard(unlockWalletPassword);
+        mUnlockWalletPassword.setText(null);
+        mOnboardingViewModel.setPassword(unlockWalletPassword);
+        if (mOnNextPage != null) {
+            mOnNextPage.incrementPages(1);
         }
     }
+
+    //    @SuppressLint("MissingPermission")
+    //    @RequiresApi(api = Build.VERSION_CODES.P)
+    //    private void showBiometricAuthenticationDialog() {
+    //        final AuthenticationCallback authenticationCallback =
+    //                new AuthenticationCallback() {
+    //                    @Override
+    //                    public void onAuthenticationSucceeded(
+    //                            AuthenticationResult authenticationResult) {
+    //                        super.onAuthenticationSucceeded(authenticationResult);
+    //
+    //                        final KeyringService keyringService = getKeyringService();
+    //                        String unlockWalletPassword;
+    //
+    //                        try {
+    //                            unlockWalletPassword = KeystoreHelper.decryptText();
+    //                        } catch (InvalidAlgorithmParameterException
+    //                                | UnrecoverableEntryException
+    //                                | NoSuchPaddingException
+    //                                | IllegalBlockSizeException
+    //                                | CertificateException
+    //                                | KeyStoreException
+    //                                | NoSuchAlgorithmException
+    //                                | BadPaddingException
+    //                                | IOException
+    //                                | InvalidKeyException e) {
+    //                            // KeystoreHelper.decryptText() may throw a long list
+    //                            // of exceptions.
+    //                            showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                            return;
+    //                        }
+    //
+    //                        if (TextUtils.isEmpty(unlockWalletPassword) || keyringService == null)
+    // {
+    //                            showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                            return;
+    //                        }
+    //
+    //                        keyringService.unlock(
+    //                                unlockWalletPassword,
+    //                                result -> {
+    //                                    if (result) {
+    //                                        Utils.clearClipboard(unlockWalletPassword);
+    //                                        mUnlockWalletPassword.setText(null);
+    //
+    // mOnboardingViewModel.setPassword(unlockWalletPassword);
+    //                                        if (mOnNextPage != null) {
+    //                                            mOnNextPage.incrementPages(1);
+    //                                        }
+    //                                    } else {
+    //
+    // showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                                    }
+    //                                });
+    //                    }
+    //
+    //                    @Override
+    //                    public void onAuthenticationError(int errorCode, CharSequence errString) {
+    //                        super.onAuthenticationError(errorCode, errString);
+    //
+    //                        final Context context = getContext();
+    //                        if (!TextUtils.isEmpty(errString) && context != null) {
+    //                            Toast.makeText(context, errString, Toast.LENGTH_SHORT).show();
+    //                        }
+    //                        showBiometricAuthenticationButton(mBiometricUnlockButton);
+    //                    }
+    //                };
+    //        Executor executor = ContextCompat.getMainExecutor(requireContext());
+    //        new Builder(requireContext())
+    //                .setTitle(getResources().getString(R.string.fingerprint_unlock))
+    //                .setDescription(getResources().getString(R.string.use_fingerprint_text))
+    //                .setNegativeButton(
+    //                        getResources().getString(android.R.string.cancel),
+    //                        executor,
+    //                        (dialog, which) ->
+    //                                authenticationCallback.onAuthenticationError(
+    //                                        BIOMETRIC_ERROR_USER_CANCELED, ""))
+    //                .build()
+    //                .authenticate(new CancellationSignal(), executor, authenticationCallback);
+    //    }
 }
